fix generateBlock in integration tests
